wl_base: Make set_content_protection poll for current status (v2)
After setting content protection to on, we should poll to see
if the compositor was able to set it or not. We poll for up to
10 Vblanks and if it still hasn't been able to set it on, then
we return false. This is a sane choice because if by 10
iterations, we still haven't seen the content protection
status change, then something is really wrong and it isn't
changing anymore. In case it was able to or the caller wanted to
turn content protection off, we return true.

v2: Explained why 10 iterations was chosen both in a commit and
in comments.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>
Reviewed-by: Daniel Charles<daniel.charles@intel.com>