gl-renderer: Only provide RBC modifiers if specifically asked for
By default, IAS never turned on RBC. However, gl-renderer was
providing RBC modifiers to linux-dmabuf which was then forwarding
them to Mesa which chose RBC buffers for apps. This would have
normally been okay, except in cases where we were using libva
to encode content. Since libva doesn't support RBC, it encoded
a compressed buffer which would look corrupted when decompressed.
gl-renderer component was at fault here because it was providing
all modifiers to linux-dmabuf. With this change, we are now
filtering out RBC modifiers in case the use_rbc flag is not set
to 1.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>