vmdisplay: Changed wl_drm version back to v2 from v3.
A private version 3 of wl_drm was introduced to add support for
creation of tiled prime buffers. Since newer versions of Mesa
now come with modifiers support, this change is no longer
required. As such, we are moving back to version 2 of wl_drm.

Note the large number of changes to wayland-drm-protocol.c is
purely because of a new version of wayland-scanner auto generating
this file.

Fixes: https://github.com/intel/ias/issues/37

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>