RemoteDisplay: add remote-display-show
This application allows us to view content that we are about to
encode and transfer to another SoC. A standalone utility called
remote-display-show will get DRM prime handles of buffers of other
apps that are being provided to the IAS compositor. This app can then
show them on the screen. Additionally, we can also do the following
things:
• Easily test remote display
• Picture in Picture, replicate the same picture multiple time.
• Size/Rotate/Scale/Crop without impacting the original app (that
can be streamed to other display)

It supports both EGL & SHM buffers.

usage:
  start weston-simple-egl and weston-simple-shm so we have
  	the 2 types (egl & shm)
then:
    remote-display-show --surfname=simple-shm --z=100 --x=100
    	--y=200 --w=640 --h=480
    remote-display-show --surfname=simple-egl --z=100 --x=300
    	--y=300 --w=800 --h=600
Show the fullscreen ouput=0 as a texture & rotate it by 180
    remote-display-show --output=0 --trsf=0.0,1.0,1.0,0.0

Signed-off-by: Philippe Lecluse <philippe.lecluse@intel.com>
Reviewed-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>