ias-shell: Print_fps for each output's flip count
Previously, we were only printing the flips per second for all
surfaces regardless of the outputs they were on. With this change
we have organized it a little better. We now show each output's
flip counts as well as which surfaces are on that output, what
their attach rate is (i.e. the rate at which apps are providing
their buffers to the compositor) as well as the rate at which
we are flipping these buffers on the screen.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>