RemoteDisplay: Fixed an issue with surface changing outputs
If a surface whose content was being sent remotely to another
SoC, then moved to a different Wayland output, there was an
issue where we would stop encoding and transmitting its buffers.
The problem was that a surface's encoding was tied to the output
that we started capturing it upon. If this surface moved to another
output and there was no other surface on the previous output, then
we would stop encoding because the output wouldn't repaint again.
Another issue was that we don't encode faster than vsync rate
to avoid overburdening VDENC. So if a surface moves to another
output, then we need to tie the capture to the vsync for this new
output.
This change fixes both of these issues.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>