gl-renderer: Turn off Y tiled buffers if RBC isn't on.
When an app wants to create a buffer, it goes to Mesa. Mesa
checks the compositor for what modifiers (tiling formats) can
be selected for this buffer. The compositor in turn goes back
to Mesa and asks for what modifiers are supportable for this
gen etc. Mesa finally selects the best modifier and creates a
color buffer.
When i915 receives this color buffer, it checks its modifier.
It has a check where for "Non Y-tile" it uses only 8 DDB blocks,
but for Y tile and CCS buffers, it calculates the DDB like this:
DIV_ROUND_UP((4 * src_w * plane_bpp), 512) * min_scanlines/4 + 3
This calculation always yields in DDBs needed for 4K planes to be
above DDB allocation on APL. As such, neither Y_TILED, nor
Y_TILED_CCS buffers are flippable on 4K monitors. The right thing
to do would be to turn off Y tiling or Y tiling CCS in case of 4K
surfaces but we don't know ahead of time how big of a surface
would be created.
Therefore, filtering out Y tiling mods from the compositor when Mesa
comes around asking for modifiers at buffer creation time. They can
still be provided by the compositor to mesa if the "use_rbc" flag is
turned on from ias.conf.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>