compositor-ias: Enable Wayland head's connection status
Starting with 5.0, there is a concept of weston heads. The
relevant changes in IAS were made in this commit:
751ddc033a808d2150493ce6fa541797be88e489

One thing that was unfortunately left out was that the head's
status must be set to connected or else weston_head_is_connected
will always return false. This will cause input devices to not
be correctly connected to output devices if we want to bind
them specifically with udev rules.
Note that with this change, if you want to use a udev rule to do
so, you must specify the ias_crtc name (not the output name) from
ias.conf in your udev rule.
For example, if ias.conf looks like this:
<crtc name='HDMI1' model='classic' mode='preferred'>
	<output name='left' size='inherit' position='origin' />
</crtc>

Then your udev rule should look something like this:
ATTRS{phys}=="usb-0000:00:15.0-3/input0", ENV{WL_OUTPUT}="HDMI1"

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>