RemoteDisplay: Made transport data private to the plugin
There was no reason to pass a private data structure from the
encoder to the transport plugin if it was truly private to the
plugin. So, removed all references of this data structure from
the encoder and made it simpler for the plugin. Also, now since
we don't have to pass it and as it is global for the rest of the
plugin, we can have other functions use it as well.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>