RemoteDisplay: Modified cmd line to provide input port with data port.
This is a fairly large change. The previous command line for remote
display required that we provide
--clients=<IP addr1>:data_port1,<IP addr2>:data_port2 to stream
frames to an IP address and a data port of multiple receivers. However,
it required that we provide the same IP address and an input port again
for receiving input events. Furthermore, there was no provision to
receive input events from multiple receivers. With this change, we
remove the need for providing the IP address all over again and change
the format such that it looks like this:
--clients=<IP addr1>:data_port1:input_port1,<IP addr2>:data_port2:input_port2

Internally, this required several key changes. One of the biggest one
was to have some communication between the transport plugin and input
receiver. Transport plugin can now export a function called
get_sockaddr. This will be called by input_receiver to get a list of
the IP addresses, and input ports that are provided by the user on the
command line. Note that this is optional and if the transport plugin
doesn't provide this function (such as avb, file..), then we can still
stream data but the input receiver will just not work.

Note that the previous cmd line arguments of relay_input_ipaddr and
relay_input_port are now gone.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>