vmdisplay: Fixed compiler warnings.
We can't strncpy upto the size of the source or destination buffers.
It has to be 1 less than the size. In vmdisplay-server-network.cpp's
case, we were copying 255 bytes when we should only be copying upto 254.
In vmdisplay-server.cpp's case, the source buffer is only 108 bytes
long, so changed the destination size to be 100 bytes only. This should
be sufficient to hold the socket path.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>