plugins: Fixed switch statement fallthrough warning
GCC7 onwards will throw a warning if it sees an implicit case
statement fallthrough to the next case statement. In our case,
we intentionally wanted it to go to the next case statement. To
prevent the compiler from throwing a warning in that situation, we
have now added  __attribute__ ((fallthrough));

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>