Role reversal from UDP client to server for input
In order to receive UDP packet, the entity has to bind to a port.
If an entity binds to a port and waits for the packet, then it is typically
called UDP server. Previously, remote-display transport plugin udp was
sending a hello message to other SoC and then would start receiving input
events packet. Now, it binds to a port and can listen to packets without
hello message. This removed start order dependency.