ias-shell: Cleanup capture proxy on application termination
If there is a remote-display instance which captures the surface
of another application, then there is a possibility that the application
gets terminated before the remote-display. In such cases, the capture
proxy of the corresponding surface is not cleaned up resulting in memory
leak. Moreover, if another application comes up and requests a surface
the compositor can assign the same surfid. In this case, the remote-display
instance to capture the newly created surface exits with failure to capture.
This commit fixes the above by cleaning up the capture proxy as well
if an application who has a surface gets terminated.

Signed-off-by: Gowtham Hosamane <gowtham.hosamane@intel.com>