dma_test: Added an app to display IPU content on screen through dma_buf.
dma_test is an application to show the capability of DMA buffers
in processing data from IPU driver. As a result, the camera
contents will be convert to ARGB, and display on the screen
through dma_buffer way using Wayland.

I am not the owner of this application, instead, I am open
sourcing it from an internal git repo.

This application includes static libraries and header files from
LinuxTV community, which is under GNU LGPL2.1 (or later) License.

Signed-off-by: Lee, Shea Qi <shea.qi.lee@intel.com>