meson.build: Made a dynamic lib from ias-shell-client-protocol
Several client apps that live outside of IAS source want to compile
their source with just a dynamic lib instead of adding the source files
in their code. With autotools way, we used to provide a dynamic library
called ias-shell-protocol.so. This was missed when we moved to meson
build. With this patch we are readding this capability to build the *.so
as well as provide header files needed by these client apps.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>