wl_base: Fixed an issue where Wayland events weren't coming to client
There were two issues here. One was that dispatch_pending is not enough
to get all the events from the server to the client. It also required a
roundtrip. The second issue was that we need to pump the socket to get
all the events from the compositor after we have sent a content protection
enable request to the compositor so need to call dispatch_pending.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>