dma_test: Removed -E flag in example.
During execution, -E flag is being used, dma_test tool will be
using dma buffer exported from the IPU driver instead of GPU
driver, and the performance is not as optimized in that case.
Higher CPU load has been observed with -E so removing it from the
readme.txt as the recommended way to be used.

Signed-off-by: Lee, Shea Qi <shea.qi.lee@intel.com>
Reviewed-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>