Set presetation clock while initialising the IAS compositor
The presentation clock for the compositor is set by weston_compositor_set_presentation_clock.
This method was missing from ias.c, thus the presentation_clock remained with the default value (0),
which is defined as CLOCK_REALTIME in time.h.

If the presentation_clock is CLOCK_REALTIME and the user changes the date continuously the image will shatter or will freeze.
In this case CLOCK_MONOTONIC is preferred.